requirements.txt
faust-streaming
SQLAlchemy
psycopg2-binary
python-dotenv
